Change Samsung Internet API support from null to true based on tests
Tests were generated by https://github.com/foolip/mdn-bcd-collector
and results collected in Samsung Internet 9 on Android 8:
https://github.com/foolip/mdn-bcd-results/pull/115

This is similar to https://github.com/mdn/browser-compat-data/pull/3700,
which has examples of what the tests look like.

Where the results were true and the existing data in BCD was null, the
value was changed to true. Snapshot of script used:
https://gist.github.com/foolip/afecccbff07ae9e0c1c788a1bb9549f1

Plain false positives are very unlikely, but some of the exposed APIs
might still need notes or could be partial implementations.