Remove entries for cloneNode() and importNode() deep parameters
These entries represent not the parameter itself, but whether it's
optional and defaulting to false. Since that is the current spec'd
behavior, these entries aren't irrelevant in the "removed for two years"
sense, but if they were inverted to mean "defaults to true" they would
be.

This issue is already documented on both MDN pages:
https://developer.mozilla.org/en-US/docs/Web/API/Node/cloneNode
https://developer.mozilla.org/en-US/docs/Web/API/Document/importNode