Add a migration script to set experimental to false
This is to bring the data in line with the recently added guideline:
https://github.com/mdn/browser-compat-data/blob/main/docs/data-guidelines.md#choosing-an-experimental-status

Lots of features which are now widely supported are still marked as
experimental. This migration script conservatively sets experimental to
false for features which are supported in Chrome+Firefox+Safari.

It could later be tweaked to do the same for 2-engine support.